Refactor tests to typescript (#178)
