+++ tests/workflows/workflows-service.test.ts
@@ -1,3 +1,5 @@
+import WorkflowStateInterface, {WorkflowStateType} from '../../src/workflow-states/contracts/workflow-state-interface'
+import Workflow from '../../src/workflows/workflow'
 import WorkflowService, {WORKFLOW_STATES} from '../../src/workflows/workflows-service'
 
 
@@ -7,26 +9,26 @@ describe('WorkflowService', () => {
   })
   it('should get workflow stats and return array', () => {
     jest.spyOn(WorkflowService.prototype, 'list').mockResolvedValue([{
-      id: 'UUID1',
+      id: 1,
       name: 'Test Workflow',
-      states: [{id: 'UUID2', name: 'Test State'}]
-    }])
+      states: [{id: 2, name: 'Test State'} as unknown as WorkflowStateInterface]
+    } as Workflow])
     const service = new WorkflowService({headers: {}})
-    const expectedStates = [{id: 'UUID2', name: 'Test State'}]
+    const expectedStates = [{id: 2, name: 'Test State'}]
     expect(service.getWorkflowStates()).resolves.toEqual(expectedStates)
   })
 
   it('should get workflow state from cached responses', async () => {
     jest.spyOn(WorkflowService.prototype, 'list').mockResolvedValue([{
-      id: 'UUID1',
+      id: 1,
       name: 'Test Workflow',
-      states: [{id: 'UUID2', name: 'Test State'}]
-    }])
+      states: [{id: 2, name: 'Test State'} as unknown as WorkflowStateInterface]
+    } as Workflow])
     WORKFLOW_STATES[123] = {
       id: 123,
       name: 'Test State',
       color: 'red',
-      type: 'Unstarted',
+      type: WorkflowStateType.UNSTARTED,
       createdAt: new Date(),
       description: 'Test Description',
       entityType: 'Story',
@@ -35,7 +37,6 @@ describe('WorkflowService', () => {
       numStories: 1,
       numStoryTemplates: 1,
       updatedAt: new Date(),
-      appUrl: 'http://test.com'
     }
     const service = new WorkflowService({headers: {}})
     await service.getWorkflowState(123)
